# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: C2S Cyber Assurance Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /controls:
    get:
      operationId: listControls
      tags: [controls]
      summary: List all canonical controls
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [effective, degraded, failed, unknown, inherited, compensating]
        - name: framework
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of controls
          content:
            application/json:
              schema:
                type: object
                properties:
                  controls:
                    type: array
                    items:
                      $ref: "#/components/schemas/Control"
                  total:
                    type: integer
                required:
                  - controls
                  - total

    post:
      operationId: createControl
      tags: [controls]
      summary: Create a canonical control
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateControlBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"
...
  /controls/{controlId}:
    get:
      operationId: getControl
      tags: [controls]
      summary: Get a control by ID
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Control
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"

    patch:
      operationId: updateControl
      tags: [controls]
      summary: Update a control
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateControlBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"

  # ── FRAMEWORKS ───────────────────────────────────────────────────────────────
...
  /frameworks/{frameworkId}/compliance:
    get:
      operationId: getFrameworkCompliance
      tags: [frameworks]
      summary: Get compliance score for a framework
      parameters:
        - name: frameworkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Compliance score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrameworkCompliance"
...
components:
  schemas:
...
    Control:
      type: object
      properties:
        id:
          type: string
        controlId:
          type: string
          description: Canonical control ID e.g. UCO-001
        name:
          type: string
        objective:
          type: string
        status:
          type: string
          enum: [effective, degraded, failed, unknown, inherited, compensating]
        severity:
          type: string
          enum: [critical, high, medium, low]
        effectiveness:
          type: number
          description: 0-100 effectiveness score
        frameworks:
          type: array
          items:
            type: string
        telemetrySources:
          type: array
          items:
            type: string
        lastValidated:
          type: string
          format: date-time
        automationCapability:
          type: string
          enum: [full, partial, manual]
        maturityLevel:
          type: integer
          minimum: 1
          maximum: 5
        driftDetected:
          type: boolean
        evidenceFresh:
          type: boolean
        remediationGuidance:
          type: string
      required:
        - id
        - controlId
        - name
        - objective
        - status
        - severity
        - effectiveness
        - frameworks
        - telemetrySources
        - lastValidated
        - automationCapability
        - maturityLevel
        - driftDetected
        - evidenceFresh
    CreateControlBody:
      type: object
      properties:
        controlId:
          type: string
        name:
          type: string
        objective:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        frameworks:
          type: array
          items:
            type: string
        automationCapability:
          type: string
          enum: [full, partial, manual]
        maturityLevel:
          type: integer
        remediationGuidance:
          type: string
      required:
        - controlId
        - name
        - objective
        - severity
        - frameworks
        - automationCapability
        - maturityLevel
    UpdateControlBody:
      type: object
      properties:
        status:
          type: string
          enum: [effective, degraded, failed, unknown, inherited, compensating]
        effectiveness:
          type: number
        remediationGuidance:
          type: string
        maturityLevel:
          type: integer
...
    FrameworkCompliance:
      type: object
      properties:
        frameworkId:
          type: string
        score:
          type: number
        controlsMapped:
          type: integer
        controlsPassing:
          type: integer
        controlsFailing:
          type: integer
        gaps:
          type: array
          items:
            type: string
        lastAssessed:
          type: string
          format: date-time
      required:
        - frameworkId
        - score
        - controlsMapped
        - controlsPassing
        - controlsFailing
        - gaps
        - lastAssessed
...